Split Debug plugin into DebugPrompt
