+++ pyboy/plugins/__init__.py
@@ -8,18 +8,19 @@ Plugins that extend PyBoy's functionality. The only publicly exposed, are the ga
 
 __pdoc__ = {
     # docs exclude
-    "window_headless": False,
-    "window_open_gl": False,
-    "screen_recorder": False,
-    "rewind": False,
-    "window_dummy": False,
-    "disable_input": False,
-    "manager_gen": False,
-    "auto_pause": False,
-    "manager": False,
-    "record_replay": False,
-    "screenshot_recorder": False,
-    "debug": False,
     "window_sdl2": False,
+    "window_open_gl": False,
+    "debug": False,
+    "window_dummy": False,
+    "auto_pause": False,
+    "rewind": False,
+    "window_headless": False,
+    "screen_recorder": False,
+    "manager_gen": False,
+    "disable_input": False,
+    "screenshot_recorder": False,
+    "debug_prompt": False,
+    "record_replay": False,
+    "manager": False,
     # docs exclude end
 }
