+++ pyboy/plugins/debug_prompt.pxd
@@ -1,0 +1,16 @@
+#
+# License: See LICENSE.md file
+# GitHub: https://github.com/Baekalfen/PyBoy
+#
+
+from pyboy.logging.logging cimport Logger
+from pyboy.plugins.base_plugin cimport PyBoyPlugin
+
+
+cdef Logger logger
+
+cdef class DebugPrompt(PyBoyPlugin):
+    cdef dict rom_symbols
+    pass
+
+
