+++ pyboy/plugins/manager.pxd
@@ -7,9 +7,8 @@ cimport cython
 
 from pyboy.logging.logging cimport Logger
 from pyboy.plugins.auto_pause cimport AutoPause
-# imports end
-from pyboy.plugins.base_plugin cimport PyBoyPlugin, PyBoyWindowPlugin
 from pyboy.plugins.debug cimport Debug
+from pyboy.plugins.debug_prompt cimport DebugPrompt
 from pyboy.plugins.disable_input cimport DisableInput
 from pyboy.plugins.game_wrapper_kirby_dream_land cimport GameWrapperKirbyDreamLand
 from pyboy.plugins.game_wrapper_pokemon_gen1 cimport GameWrapperPokemonGen1
@@ -22,6 +21,8 @@ from pyboy.plugins.screenshot_recorder cimport ScreenshotRecorder
 from pyboy.plugins.window_dummy cimport WindowDummy
 from pyboy.plugins.window_headless cimport WindowHeadless
 from pyboy.plugins.window_open_gl cimport WindowOpenGL
+# imports end
+# isort:skip
 # imports
 from pyboy.plugins.window_sdl2 cimport WindowSDL2
 
@@ -44,6 +45,7 @@ cdef class PluginManager:
     cdef public Rewind rewind
     cdef public ScreenRecorder screen_recorder
     cdef public ScreenshotRecorder screenshot_recorder
+    cdef public DebugPrompt debug_prompt
     cdef public GameWrapperSuperMarioLand game_wrapper_super_mario_land
     cdef public GameWrapperTetris game_wrapper_tetris
     cdef public GameWrapperKirbyDreamLand game_wrapper_kirby_dream_land
@@ -59,6 +61,7 @@ cdef class PluginManager:
     cdef bint rewind_enabled
     cdef bint screen_recorder_enabled
     cdef bint screenshot_recorder_enabled
+    cdef bint debug_prompt_enabled
     cdef bint game_wrapper_super_mario_land_enabled
     cdef bint game_wrapper_tetris_enabled
     cdef bint game_wrapper_kirby_dream_land_enabled
